chore(webpack):  Split webpack config and setup dll plugin to improve performance;
